requirements.txt
alembic==1.7.6
Babel==2.9.1
bidict==0.22.0
blinker==1.4
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
click==7.1.2
colorama==0.4.4
cryptography==37.0.2
dominate==2.6.0
Flask==1.1.4
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
flask-paginate==2022.1.8
Flask-SocketIO==5.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
google-auth==2.6.5
greenlet==1.1.2
importlib-metadata==4.11.3
importlib-resources==5.7.1
itsdangerous==0.24
Jinja2==2.11.3
Mako==1.1.6
MarkupSafe==2.0.1
prettytable==3.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyecharts==1.9.1
pyOpenSSL==22.0.0
python-engineio==4.3.2
python-socketio==5.6.0
pytz==2022.1
rsa==4.8
simplejson==3.17.6
six==1.16.0
SQLAlchemy==1.4.32
visitor==0.1.3
wcwidth==0.2.5
Werkzeug==0.16.1
wincertstore==0.2
WTForms==2.2.1
zipp==3.7.0
gunicorn==20.1.0
gevent==21.12.0
gevent-websocket==0.10.1
